requirements.txt
Flask==2.3.3
openpyxl==3.1.2
python-dateutil==2.8.2
Werkzeug==2.3.7
jpholiday==1.0.2
requests==2.31.0
google-cloud-firestore==2.16.0
firebase-admin==6.5.0 